requirements.txt
numpy
pandas
torch
torchvision
hydra-core
mlflow
jupyter
scikit-learn
tqdm
kaggle
